refactor: move fee input to review step - vote
