+++ src/domains/transaction/pages/SendVote/ReviewStep.tsx
@@ -1,27 +1,39 @@
-import React, { useEffect } from "react";
+import React, { useEffect, useMemo } from "react";
 import { useFormContext } from "react-hook-form";
 import { useTranslation } from "react-i18next";
 import { SendVoteStepProperties } from "./SendVote.contracts";
-import { TotalAmountBox } from "@/domains/transaction/components/TotalAmountBox";
 import { StepHeader } from "@/app/components/StepHeader";
 import { DetailTitle, DetailWrapper } from "@/app/components/DetailWrapper";
 import { Address } from "@/app/components/Address";
 import { ThemeIcon } from "@/app/components/Icon";
 import { VoteTransactionType } from "@/domains/transaction/components/VoteTransactionType";
 import cn from "classnames";
-import { calculateGasFee } from "@/domains/transaction/components/InputFee/InputFee";
+import { FormField, FormLabel } from "@/app/components/Form";
+import { FeeField } from "@/domains/transaction/components/FeeField";
 
-export const ReviewStep = ({ unvotes, votes, wallet }: SendVoteStepProperties) => {
+export const ReviewStep = ({ unvotes, votes, wallet, profile }: SendVoteStepProperties) => {
 	const { t } = useTranslation();
-	const { getValues, unregister } = useFormContext();
-
-	const { gasPrice, gasLimit } = getValues();
-
-	const fee = calculateGasFee(gasPrice, gasLimit);
+	const { unregister } = useFormContext();
 
 	useEffect(() => {
 		unregister("mnemonic");
 	}, [unregister]);
+
+	const showFeeInput = useMemo(() => !wallet.network().chargesZeroFees(), [wallet]);
+
+	const feeTransactionData = useMemo(
+		() => ({
+			unvotes: unvotes.map((vote) => ({
+				amount: vote.amount,
+				id: vote.wallet?.governanceIdentifier(),
+			})),
+			votes: votes.map((vote) => ({
+				amount: vote.amount,
+				id: vote.wallet?.governanceIdentifier(),
+			})),
+		}),
+		[unvotes, votes, wallet],
+	);
 
 	return (
 		<section data-testid="SendVote__review-step">
@@ -51,13 +63,12 @@ export const ReviewStep = ({ unvotes, votes, wallet }: SendVoteStepProperties) =
 
 				<div data-testid="DetailWrapper">
 					<div className="mt-0 p-3 sm:mt-2 sm:p-0">
-						<TotalAmountBox
-							amount={0}
-							fee={fee}
-							ticker={wallet.currency()}
-							convertValues={!wallet.network().isTest()}
-							hideAmount
-						/>
+						{showFeeInput && (
+							<FormField name="fee" className="flex-1">
+								<FormLabel label={t("TRANSACTION.TRANSACTION_FEE")} />
+								<FeeField type="vote" data={feeTransactionData} network={wallet.network()} profile={profile} />
+							</FormField>
+						)}
 					</div>
 				</div>
 			</div>
